requirements.txt
gluon==1.1.0
gluoncv==0.10.5
h5py==2.10.0
numpy==1.22.3
pandas==1.3.5
Pillow==9.0.1
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==2.8.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.63.0
